standalone class for movie scrapping
